Revert "Rename settings mock"
- Avoid issues when backporting
