+++ public/app/plugins/datasource/azuremonitor/__mocks__/instanceSettings.ts
@@ -1,6 +1,6 @@
-import { DataSourceInstanceSettings, KeyValue, PluginType } from '@grafana/data';
+import { DataSourceInstanceSettings, PluginType } from '@grafana/data';
 
-import { AzureDataSourceInstanceSettings, AzureDataSourceSettings } from '../types';
+import { AzureDataSourceInstanceSettings } from '../types';
 
 import { DeepPartial, mapPartialArrayObject } from './utils';
 
@@ -67,40 +67,5 @@ export const createMockInstanceSetttings = (
       subscriptionId: 'ghi-789',
       ...overrides?.jsonData,
     },
-  };
-};
-
-export const createMockDatasourceSettings = (
-  overrides?: DeepPartial<AzureDataSourceSettings>,
-  secureJsonFieldsOverrides?: KeyValue<boolean>
-): AzureDataSourceSettings => {
-  return {
-    id: 1,
-    uid: 'uid',
-    orgId: 1,
-    name: 'test-data-source',
-    typeLogoUrl: 'logo',
-    type: 'grafana-azure-monitor-datasource',
-    typeName: 'datasource',
-    access: '',
-    url: '',
-    user: '',
-    database: '',
-    basicAuth: false,
-    basicAuthUser: '',
-    isDefault: false,
-    jsonData: {
-      cloudName: 'azuremonitor',
-      azureAuthType: 'clientsecret',
-
-      tenantId: 'abc-123',
-      clientId: 'def-456',
-      subscriptionId: 'ghi-789',
-      ...overrides?.jsonData,
-    },
-    secureJsonData: { ...overrides?.secureJsonData },
-    secureJsonFields: { ...secureJsonFieldsOverrides },
-    readOnly: false,
-    withCredentials: false,
   };
 };
