+++ public/app/plugins/datasource/azuremonitor/components/ConfigEditor/BasicLogsToggle.test.tsx
@@ -1,6 +1,6 @@
 import { render, screen } from '@testing-library/react';
 
-import { createMockInstanceSetttings } from '../../__mocks__/settings';
+import { createMockInstanceSetttings } from '../../__mocks__/instanceSettings';
 
 import { BasicLogsToggle, Props } from './BasicLogsToggle';
